Separate ci and publish.
Since the published happened after the build and test but on a separate job, it actually lacked the build artifact and published an empty package.

Here, I separate the ci executed on pull requests, and the ci executed on release, and publish the package after build and test only on release.